Clippy fixes
